+++ lib/shared/types.ts
@@ -191,15 +191,15 @@ export interface OutputFile {
   readonly text: string
 }
 
-export interface BuildResult<SpecificOptions extends BuildOptions = BuildOptions> {
+export interface BuildResult<ProvidedOptions extends BuildOptions = BuildOptions> {
   errors: Message[]
   warnings: Message[]
   /** Only when "write: false" */
-  outputFiles: OutputFile[] | (SpecificOptions['write'] extends false ? never : undefined)
+  outputFiles: OutputFile[] | (ProvidedOptions['write'] extends false ? never : undefined)
   /** Only when "metafile: true" */
-  metafile: Metafile | (SpecificOptions['metafile'] extends true ? never : undefined)
+  metafile: Metafile | (ProvidedOptions['metafile'] extends true ? never : undefined)
   /** Only when "mangleCache" is present */
-  mangleCache: Record<string, string | false> | (SpecificOptions['mangleCache'] extends Object ? never : undefined)
+  mangleCache: Record<string, string | false> | (ProvidedOptions['mangleCache'] extends Object ? never : undefined)
 }
 
 export interface BuildFailure extends Error {
@@ -253,14 +253,14 @@ export interface TransformOptions extends CommonOptions {
   footer?: string
 }
 
-export interface TransformResult<SpecificOptions extends TransformOptions = TransformOptions> {
+export interface TransformResult<ProvidedOptions extends TransformOptions = TransformOptions> {
   code: string
   map: string
   warnings: Message[]
   /** Only when "mangleCache" is present */
-  mangleCache: Record<string, string | false> | (SpecificOptions['mangleCache'] extends Object ? never : undefined)
+  mangleCache: Record<string, string | false> | (ProvidedOptions['mangleCache'] extends Object ? never : undefined)
   /** Only when "legalComments" is "external" */
-  legalComments: string | (SpecificOptions['legalComments'] extends 'external' ? never : undefined)
+  legalComments: string | (ProvidedOptions['legalComments'] extends 'external' ? never : undefined)
 }
 
 export interface TransformFailure extends Error {
@@ -487,9 +487,9 @@ export interface AnalyzeMetafileOptions {
 export interface WatchOptions {
 }
 
-export interface BuildContext<SpecificOptions extends BuildOptions = BuildOptions> {
+export interface BuildContext<ProvidedOptions extends BuildOptions = BuildOptions> {
   /** Documentation: https://esbuild.github.io/api/#rebuild */
-  rebuild(): Promise<BuildResult<SpecificOptions>>
+  rebuild(): Promise<BuildResult<ProvidedOptions>>
 
   /** Documentation: https://esbuild.github.io/api/#watch */
   watch(options?: WatchOptions): Promise<void>
@@ -501,6 +501,11 @@ export interface BuildContext<SpecificOptions extends BuildOptions = BuildOption
   dispose(): Promise<void>
 }
 
+// This is a TypeScript type-level function which replaces any keys in "In"
+// that aren't in "Out" with "never". We use this to reject properties with
+// typos in object literals. See: https://stackoverflow.com/questions/49580725
+type SameShape<Out, In extends Out> = In & { [Key in Exclude<keyof In, keyof Out>]: never }
+
 /**
  * This function invokes the "esbuild" command-line tool for you. It returns a
  * promise that either resolves with a "BuildResult" object or rejects with a
@@ -511,8 +516,7 @@ export interface BuildContext<SpecificOptions extends BuildOptions = BuildOption
  *
  * Documentation: https://esbuild.github.io/api/#build
  */
-export declare function build<SpecificOptions extends BuildOptions>(options: SpecificOptions): Promise<BuildResult<SpecificOptions>>
-export declare function build(options: BuildOptions): Promise<BuildResult>
+export declare function build<T extends BuildOptions>(options: SameShape<BuildOptions, T>): Promise<BuildResult<T>>
 
 /**
  * This is the advanced long-running form of "build" that supports additional
@@ -523,8 +527,7 @@ export declare function build(options: BuildOptions): Promise<BuildResult>
  *
  * Documentation: https://esbuild.github.io/api/#build
  */
-export declare function context<T extends BuildOptions>(options: T): Promise<BuildContext<T>>
-export declare function context(options: BuildOptions): Promise<BuildContext>
+export declare function context<T extends BuildOptions>(options: SameShape<BuildOptions, T>): Promise<BuildContext<T>>
 
 /**
  * This function transforms a single JavaScript file. It can be used to minify
@@ -537,8 +540,7 @@ export declare function context(options: BuildOptions): Promise<BuildContext>
  *
  * Documentation: https://esbuild.github.io/api/#transform
  */
-export declare function transform<SpecificOptions extends TransformOptions>(input: string | Uint8Array, options?: SpecificOptions): Promise<TransformResult<SpecificOptions>>
-export declare function transform(input: string | Uint8Array, options?: TransformOptions): Promise<TransformResult>
+export declare function transform<T extends TransformOptions>(input: string | Uint8Array, options?: SameShape<TransformOptions, T>): Promise<TransformResult<T>>
 
 /**
  * Converts log messages to formatted message strings suitable for printing in
@@ -570,8 +572,7 @@ export declare function analyzeMetafile(metafile: Metafile | string, options?: A
  *
  * Documentation: https://esbuild.github.io/api/#build
  */
-export declare function buildSync<SpecificOptions extends BuildOptions>(options: SpecificOptions): BuildResult<SpecificOptions>
-export declare function buildSync(options: BuildOptions): BuildResult
+export declare function buildSync<T extends BuildOptions>(options: SameShape<BuildOptions, T>): BuildResult<T>
 
 /**
  * A synchronous version of "transform".
@@ -581,8 +582,7 @@ export declare function buildSync(options: BuildOptions): BuildResult
  *
  * Documentation: https://esbuild.github.io/api/#transform
  */
-export declare function transformSync<SpecificOptions extends TransformOptions>(input: string, options?: SpecificOptions): TransformResult<SpecificOptions>
-export declare function transformSync(input: string | Uint8Array, options?: TransformOptions): TransformResult
+export declare function transformSync<T extends TransformOptions>(input: string | Uint8Array, options?: SameShape<TransformOptions, T>): TransformResult<T>
 
 /**
  * A synchronous version of "formatMessages".